Add files via upload
This is a fictional restaurant chain in France opened in 2025. Data was generated with ChatGPT. This infographic shows insights directly in the report.

To create this report I used my knowledge of PowerBI. For this project, I used data cleaning, data modeling, DAX measures, creation of visuals adapted to insight needed, report interaction, and storytelling.